refactor: change package structure in data
- divide database and remote package
