+++ app/src/main/java/woowacourse/shopping/data/database/cart/CartDao.kt
@@ -3,7 +3,7 @@ package woowacourse.shopping.data.database.cart
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.Query
-import woowacourse.shopping.data.model.entity.CartItemEntity
+import woowacourse.shopping.data.database.entity.CartItemEntity
 
 @Dao
 interface CartDao {
