+++ app/src/main/java/woowacourse/shopping/data/database/entity/CartItemEntity.kt
@@ -1,23 +1,25 @@
-package woowacourse.shopping.data.model.entity
+package woowacourse.shopping.data.database.entity
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import woowacourse.shopping.domain.model.Product
 
-@Entity(tableName = "recent_products")
-data class RecentProductEntity(
+@Entity(tableName = "cart_items")
+data class CartItemEntity(
     @PrimaryKey(autoGenerate = true) val id: Long = 0L,
     val productId: Long,
     val productName: String,
+    val price: Long,
     val imgUrl: String,
-    val category: String,
+    val quantity: Int,
 )
 
-fun Product.mapper(): RecentProductEntity {
-    return RecentProductEntity(
+fun Product.mapper(quantity: Int): CartItemEntity {
+    return CartItemEntity(
         productId = this.id,
         productName = this.name,
+        price = this.price,
         imgUrl = this.imageUrl,
-        category = this.category,
+        quantity = quantity,
     )
 }
