+++ app/src/main/java/woowacourse/shopping/data/database/entity/RecentProductEntity.kt
@@ -1,25 +1,23 @@
-package woowacourse.shopping.data.model.entity
+package woowacourse.shopping.data.database.entity
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import woowacourse.shopping.domain.model.Product
 
-@Entity(tableName = "cart_items")
-data class CartItemEntity(
+@Entity(tableName = "recent_products")
+data class RecentProductEntity(
     @PrimaryKey(autoGenerate = true) val id: Long = 0L,
     val productId: Long,
     val productName: String,
-    val price: Long,
     val imgUrl: String,
-    val quantity: Int,
+    val category: String,
 )
 
-fun Product.mapper(quantity: Int): CartItemEntity {
-    return CartItemEntity(
+fun Product.mapper(): RecentProductEntity {
+    return RecentProductEntity(
         productId = this.id,
         productName = this.name,
-        price = this.price,
         imgUrl = this.imageUrl,
-        quantity = quantity,
+        category = this.category,
     )
 }
