+++ app/src/main/java/woowacourse/shopping/data/mapper/RecentProductMapper.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.data.mapper
 
-import woowacourse.shopping.data.model.entity.RecentProductEntity
+import woowacourse.shopping.data.database.entity.RecentProductEntity
 import woowacourse.shopping.domain.model.RecentProduct
 
 fun RecentProductEntity.toDomainModel(): RecentProduct {
