+++ app/src/main/java/woowacourse/shopping/data/repository/RecentProductRepositoryImpl.kt
@@ -3,7 +3,7 @@ package woowacourse.shopping.data.repository
 import android.content.Context
 import woowacourse.shopping.data.database.ShoppingDatabase
 import woowacourse.shopping.data.mapper.toDomainModel
-import woowacourse.shopping.data.model.entity.mapper
+import woowacourse.shopping.data.database.entity.mapper
 import woowacourse.shopping.domain.model.Product
 import woowacourse.shopping.domain.model.RecentProduct
 import woowacourse.shopping.domain.repository.RecentProductRepository
