+++ app/src/main/java/woowacourse/shopping/presentation/ui/payment/PaymentViewModel.kt
@@ -6,7 +6,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.launch
-import woowacourse.shopping.data.database.OrderDatabase
+import woowacourse.shopping.data.database.order.OrderDatabase
 import woowacourse.shopping.domain.model.Order
 import woowacourse.shopping.domain.model.coupon.Coupon
 import woowacourse.shopping.domain.repository.OrderRepository
